requirements.txt
streamlit
google.generativeai
python-docx
tiktoken
oracledb
python-dotenv
langchain_ollama

